refactor(api): split createUser from user service
