+++ packages/api/tests/modules/user/repositories/user.repository.test.ts
@@ -5,6 +5,7 @@ import { ObjectId } from "mongodb";
 import userRepository from "../../../../src/modules/user/repositories/user.repository";
 import userService from "../../../../src/modules/user/user.service";
 import { UserNotPersisted } from "../../../../src/modules/user/repositories/dbo/UserDbo";
+import userCrudService from "../../../../src/modules/user/services/crud/user.crud.service";
 
 describe("UserRepository", () => {
     const defaultUser = {
@@ -19,7 +20,7 @@ describe("UserRepository", () => {
     };
     describe("The methods must not return any secret", () => {
         beforeEach(async () => {
-            await userService.createUser({ email: "test@beta.gouv.fr" });
+            await userCrudService.createUser({ email: "test@beta.gouv.fr" });
         });
 
         it("findByEmail", async () => {
@@ -54,7 +55,7 @@ describe("UserRepository", () => {
 
     describe("getUserWithSecretsByEmail", () => {
         beforeEach(async () => {
-            await userService.createUser({ email: "test@beta.gouv.fr" });
+            await userCrudService.createUser({ email: "test@beta.gouv.fr" });
         });
 
         it("should return user", async () => {
